clustering/requirements.txt
matplotlib>=3.3.0
pandas>=1.1.5
py4j==0.10.9
pyspark-dbscan
scikit-learn
scipy
seaborn==0.11.2
sklearn
